refactor: 데이터를 받아오는 곳에서 parse로 데이터 처리, 관련 타입 수정
