+++ src/types/model.ts
@@ -40,6 +40,7 @@ interface Post {
   author: User;
   createdAt: string;
   updatedAt: string;
+  content?: string;
 }
 
 interface Like {
